prod/deny_policy_creation_service/requirements.txt
google-cloud-tasks
Jinja2
PyYAML
google-cloud-bigquery
google-auth
google-cloud
